refac(circom): move `PrimeField` to `circomlib/base`
