+++ vendors/circom/circomlib/base/prime_field.h
@@ -1,0 +1,43 @@
+#ifndef VENDORS_CIRCOM_CIRCOMLIB_BASE_PRIME_FIELD_H_
+#define VENDORS_CIRCOM_CIRCOMLIB_BASE_PRIME_FIELD_H_
+
+#include <stddef.h>
+#include <stdint.h>
+
+#include <array>
+#include <string>
+#include <vector>
+
+#include "tachyon/base/logging.h"
+#include "tachyon/math/base/big_int.h"
+
+namespace tachyon::circom {
+
+struct PrimeField {
+  std::vector<uint8_t> bytes;
+
+  bool operator==(const PrimeField& other) const {
+    return bytes == other.bytes;
+  }
+  bool operator!=(const PrimeField& other) const {
+    return bytes != other.bytes;
+  }
+
+  template <size_t N>
+  math::BigInt<N> ToBigInt() const {
+    CHECK_EQ(bytes.size() / 8, N);
+    return math::BigInt<N>::FromBytesLE(bytes);
+  }
+
+  template <size_t N>
+  static PrimeField FromBigInt(const math::BigInt<N>& big_int) {
+    std::array<uint8_t, N * 8> bytes = big_int.ToBytesLE();
+    return {{bytes.begin(), bytes.end()}};
+  }
+
+  std::string ToString() const;
+};
+
+}  // namespace tachyon::circom
+
+#endif  // VENDORS_CIRCOM_CIRCOMLIB_BASE_PRIME_FIELD_H_
